+++ railties/test/commands/server_test.rb
@@ -44,10 +44,6 @@ class Rails::Command::ServerCommandTest < ActiveSupport::TestCase
     assert_no_match(/Maybe you meant/, output)
   end
 
-  def test_using_positional_argument_deprecation
-    assert_match(/DEPRECATION WARNING/, run_command("tin"))
-  end
-
   def test_using_known_server_that_isnt_in_the_gemfile
     assert_match(/Could not load server "unicorn". Maybe you need to the add it to the Gemfile/, run_command("-u", "unicorn"))
   end